refactor data frame handling
