+++ requirements.txt
@@ -1,0 +1,2 @@
+dash==1.19.0
+pandas==1.4.4
