+++ src/version.py
@@ -1,0 +1,7 @@
+__author__ = "William Rivera"
+__credits__ = ["William Rivera", "Kyle Maclea", "Raghava Adusimilli", "Stephen Dunn"]
+__license__ = "MIT"
+__version__ = "0.1.1"
+__maintainer__ = "William Rivera"
+__email__ = "wto3@wildcats.unh.edu"
+__status__ = "Development"
